[NEW] added option to list locally available recipes and install particular recipe

+++ main/create.js
@@ -25,34 +25,48 @@ class createSite
     createLocal(type)
     {
 		var self = this;
-						
-		//createdir contains destination dir
-        if (!fs.existsSync(global.pconfig.createdir)) 
-		{         
-			fs.mkdirSync(global.pconfig.createdir);
-        }
-
-		global.pconfig.createdir = global.pconfig.createdir.replace(/\/$/, '');
-
-		fs.mkdirSync(global.pconfig.createdir+"/prevjs-site");
 		
-		var fpath = global.pconfig.createdir+"/prevjs-site";
-		
-		if (!fs.existsSync(global.pconfig.createdir)) {
-         
-			console.log("Error! Path does not exist. Try again");
-			process.exit();
-
-        }
-
 		//srcDataDir contains path to recipe in local recipe store
 		var srcDataDir = path.join(__dirname, "../recipes/"+type);
-
-		fs.copySync(srcDataDir, fpath, { overwrite: true });
 		
- 		console.log("Created at "+fpath);
-		console.log("To preview in browser type prevjs --run "+fpath+"/recipe.json");
-        process.exit();
+		if (fs.existsSync(srcDataDir)) 
+		{
+						
+			//createdir contains destination dir
+	        if (!fs.existsSync(global.pconfig.createdir)) 
+			{         
+				fs.mkdirSync(global.pconfig.createdir);
+	        }
+	
+	
+			global.pconfig.createdir = global.pconfig.createdir.replace(/\/$/, '');
+	
+			fs.mkdirSync(global.pconfig.createdir+"/prevjs-site");
+			
+			var fpath = global.pconfig.createdir+"/prevjs-site";
+			
+			if (!fs.existsSync(global.pconfig.createdir)) {
+	         
+				console.log("Error! Path does not exist. Try again");
+				process.exit();
+	
+	        }
+
+		
+	
+			fs.copySync(srcDataDir, fpath, { overwrite: true });
+			
+	 		console.log("Created at "+fpath);
+			console.log("To preview in browser type prevjs --run "+fpath+"/recipe.json");
+	        process.exit();
+	   }
+	   else
+	   {
+		
+			console.log("Recipe does not exist - " + type);
+			process.exit();
+		}
+	
 
 
     }
